requirements.txt
torch
torchvision
clip-by-openai
tqdm
numpy
